Fix confusing variable names in gtid.cc.
All the modified functions have variable "s" as parameter of the function.
And also they have local variable "s" in the "for" loop, which is even
defined through an expression on the function parameter "s". Turns out
this is not only confusing and not readable for humans, it also makes
older compilers to choke and to not understand which "s" is which. I'm
actually very surprised that modern compilers are able to parse this the
way it was intended.

I'm changing the variable names in "for" loops to be more readable and to
be different than the names of function parameters.

PiperOrigin-RevId: 256997702
Change-Id: I3c8ee1ca63e492df7267354391b27aa6fc862adf