Update all Ripple dependencies to the latest version.
On the way I'm also changing the BUILD files to comply with the latest
bazel recommendations to not rely on native cc_* and proto_* rules and
import them from bzl libraries instead. I'm doing this because new
versions of many dependencies did this change themselves, and since I
needed to adjust the WORKSPACE file to accommodate that, I've decided
to do the same modification to Ripple BUILD files.

One more extra modification: since I've discovered that
com_google_protobuf library pulls in its own version of zlib library
(that's why I had to add a call to protobuf_deps()), I've changed
Ripple to depend on the same zlib library instead of depending on
the one installed on the system. I did this just as a precaution
to not link accidentally with two different versions of zlib.

PiperOrigin-RevId: 267867014
Change-Id: Ied6612293b78b5efa74b18d786a521a0c79d28c0