Fix Ripple to always parse MySQL GTID set as having gaps.
Currently the GTIDList::Parse() method tries to be smart and set different
modes depending on what it parsed. But that has several problems in it:

1. The code assumes that if the last part in the list (GTID range with the
last uuid) doesn't have gaps, then the whole GTID set will not have
gaps (if it has, then parsing fails here
https://github.com/google/mysql-ripple/blob/master/gtid.cc#L786).

2. The code assumes that if GTIDList initially was parsed without any
gaps, then in the future MySQL master will not create any gaps either
(otherwise Ripple will fail to replicate due to validation here
https://github.com/google/mysql-ripple/blob/master/gtid.cc#L486).

I'm not sure why the code was written this way, but I think it was an
attempt of generalization that went too far. The comment in the header
file (https://github.com/google/mysql-ripple/blob/master/gtid.h#L277)
seem to confirm that the initial idea was that the mode for MySQL will
be MODE_GAPS and for MariaDB -- MODE_STRICT_MONOTONIC
(MODE_MONOTONIC was not actually implemented). I'm changing the
code to do just that -- always set the mode to MODE_GAPS for MySQL
and set it to MODE_STRICT_MONOTONIC for MariaDB.

Fixes: https://github.com/google/mysql-ripple/issues/17
PiperOrigin-RevId: 273294067
Change-Id: I67dcff0010337142df65395334c5434091e3af89